refactoring controllers and create new controller ConnectController and new model ConnectManager
