+++ modeles/CommentsManager.php
@@ -11,11 +11,11 @@ class CommentsManager extends Manager{
     public function getComments($idArticle)//méthode de récupération des commentaire avec une jointure dans la requete pour récupérer le pseudo de l'user
 	{
 		$db = $this->dbConnect();
-		$comments = $db->prepare('SELECT comments.id, users.pseudo, comments.content, comments.valid, DATE_FORMAT(comment_date, \'%d/%m/%Y à %Hh%imin%ss\') AS comment_date_fr FROM comments INNER JOIN users ON comments.id_user = users.id WHERE id_article = ? ORDER BY comment_date DESC');
+		$comments = $db->prepare('SELECT comments.id, comments.id_article, users.pseudo, comments.content, comments.valid, DATE_FORMAT(comment_date, \'%d/%m/%Y à %Hh%imin%ss\') AS comment_date_fr FROM comments INNER JOIN users ON comments.id_user = users.id WHERE id_article  = ? ORDER BY comment_date DESC');
 		$comments->execute(array($idArticle));
-		$comment = $comments->fetch();
+		//$comment = $comments->fetch();
 		//var_dump($comment); die;
-		return $comment;
+		return $comments;
     }
     
     public function postComment($idArticle, $idUser, $content)//insertion des commentaires dans la table avis
