+++ modeles/MembersManager.php
@@ -21,7 +21,6 @@ class MembersManager extends Manager{
 
     }
 
-    
     public function testMail($mail) //test pour contrer doublon mail
     
     {
@@ -32,33 +31,6 @@ class MembersManager extends Manager{
         ));
         $mailexist = $reqmail->rowCount();
         return $mailexist;
-    }
-
-    public function getConnect($mail) //récupère les information relative à la connexion de l'utilisateur inscrit en db
-    
-    {
-        $db = $this->dbConnect();
-        $req = $db->prepare('SELECT * FROM users WHERE mail =  :mail');
-        $req->execute(array(
-            'mail' => $mail
-        
-        ));
-        $connect = $req->fetch();
-        return $connect;
-    }
-
-    public function remember($mail, $mdp) // fonction se souvenir de moi
-    
-    {
-        $db = $this->dbConnect();
-        $requser = $db->prepare("SELECT * FROM users WHERE mail = ? AND pass = ?");
-        $requser->execute(array(
-            $_COOKIE['mail'],
-            $_COOKIE['pass']
-        ));
-        $usercook = $requser->rowCount();
-        return $usercook;
-
     }
 
     public function getUsers()
