+++ views/templates/admin/listUsers.html.twig
@@ -5,7 +5,7 @@
 {% block body %}
 
 <div class="articlewrite">
-    <h2>Bonjour cher admin</h2>
+    
     <br/>
     <h3><a href="index.php?action=homePage">Accueil</a></h3>
     <h3><a href="index.php?action=listArticlesAdmin">Articles</a></h3>
