+++ views/templates/articles/article.html.twig
@@ -30,8 +30,6 @@
             </div>
     </div>
 </div>
-
-
 <div align="center">
     <br>
     <div class="encart" >
@@ -48,18 +46,17 @@
        </form>
        {% else %}
           <h3 class="error">Pour l'ajout d'un commentaire, veuillez vous connecter !</h3>
-          {% endif %} 
-          
-          <div class="container">
+    {% endif %} 
+    <div class="container">
           <h2>Vos commentaires :</h2>
-          {#{% for comment in comments %}#}
-          <pre>
-            {{ dump(comment) }}
-           </pre>
           <div class="row justify-content-center text-center">
             <div class="col-6"><br>
                     <div class=" card mb-4">
                         <div class="card-body">
+                            {% for comment in comments %}
+                            <pre>
+                                {{ dump(comment) }}
+                               </pre>
                             <p><em>Envoyé le : </em>{{comment.comment_date_fr}}</p>
                             <p class="card-text" >{{comment.content|striptags|raw|nl2br}}</p></p>
                             <p class="card-text"><em>De la part de : </em>{{comment.pseudo}}</p>
@@ -67,7 +64,7 @@
                     </div>
                 </div>
             </div>
-        {#{% endfor %}#}
+        {% endfor %}
         </div>
     </div>
 </div>
