+++ views/templates/articles/articles.html.twig
@@ -11,7 +11,7 @@
     
 </div>
 <div class="container">
-    <h2>Derniers articles</h2><br>
+    <h2>Mes articles</h2><br>
     <div class="row flex">
         {% for article in articles %}
        {#<pre>
